add connection between database

+++ src/main/java/com/example/sweater/repos/MessageRepo.java
@@ -1,11 +1,12 @@
 package com.example.sweater.repos;
 
-import java.util.List;
-
 import com.example.sweater.domain.Message;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepo  extends CrudRepository<Message, Integer> {
+import java.util.List;
+
+public interface MessageRepo extends CrudRepository<Message, Long> {
 
     List<Message> findByTag(String tag);
+
 }
